# server/docs/swagger.yaml
basePath: /
definitions:
...
  types.Expense:
    properties:
      amount_baisa:
        type: integer
      category:
        $ref: '#/definitions/types.ExpenseCategory'
      created_at:
        type: string
      description:
        type: string
      group_id:
        type: string
      id:
        type: string
      occurred_on:
        type: string
      paid_by:
        type: string
      shares:
        items:
          $ref: '#/definitions/types.ShareInput'
        type: array
      split_type:
        $ref: '#/definitions/types.SplitType'
    type: object
  types.ExpenseCategory:
    enum:
    - lodging
    - fuel
    - food
    - transport
    - other
    type: string
    x-enum-varnames:
    - CategoryLodging
    - CategoryFuel
    - CategoryFood
    - CategoryTransport
    - CategoryOther
...
  types.Page:
    properties:
      items: {}
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  types.RecordExpenseRequest:
    properties:
      amount_baisa:
        type: integer
      category:
        $ref: '#/definitions/types.ExpenseCategory'
      description:
        type: string
      occurred_on:
        type: string
      shares:
        description: |-
          Shares is optional: absent/empty means an equal split across all
          approved members at settlement. Weight defaults to 1 when omitted.
        items:
          $ref: '#/definitions/types.ShareInput'
        type: array
    type: object
...
  types.ShareInput:
    properties:
      user_id:
        type: string
      weight:
        type: integer
    type: object
  types.SplitType:
    enum:
    - equal
    - subset
    - weighted
    type: string
    x-enum-varnames:
    - SplitEqual
    - SplitSubset
    - SplitWeighted
...
  types.UpdateExpenseRequest:
    properties:
      amount_baisa:
        type: integer
      category:
        $ref: '#/definitions/types.ExpenseCategory'
      description:
        type: string
      occurred_on:
        type: string
      shares:
        description: |-
          Shares is optional: absent/empty means an equal split across all
          approved members at settlement. Weight defaults to 1 when omitted.
        items:
          $ref: '#/definitions/types.ShareInput'
        type: array
    type: object
...
  types.apiError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Shared-expense settlement tracker for group trips. Members record what they paid; closing a group computes each member's fair share and the minimum set of payments that squares everyone up. Money moves off-platform — each computed payment is tracked through a two-key confirmation flow (debtor proves, creditor attests, admin finalizes).

    All amounts are integer **baisa** (1.000 OMR = 1000). There are no floats in the money path.

    **Authenticating here:** call `POST /auth/login` (try `admin@expense-splitter.local` / `admin`), copy the `access_token` from the response, click **Authorize** above and paste `Bearer <access_token>`.
  title: Expense Splitter API
  version: "1.0"
paths:
...
  /groups/{id}/expenses:
    get:
      description: |-
        Members and admins only. Filter by category or payer, and search descriptions with `q`. Returns a `{total, limit, offset, items}` envelope where `items` is an array of expenses.

        Descriptions are truncated to 80 characters, cut on a word boundary so the last word stays whole (e.g. `"dinner at the ..."`).
      parameters:
      - description: group id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: filter by category
        enum:
        - lodging
        - fuel
        - food
        - transport
        - other
        in: query
        name: category
        type: string
      - description: filter by payer's user id
        format: uuid
        in: query
        name: paid_by
        type: string
      - description: search in description
        in: query
        name: q
        type: string
      - default: 50
        description: page size (1-200)
        in: query
        name: limit
        type: integer
      - default: 0
        description: rows to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: paginated expenses
          schema:
            $ref: '#/definitions/types.Page'
        "400":
          description: invalid id, filter, or paging
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: not a member of this group
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: group not found
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: List a group's expenses
      tags:
      - expenses
    post:
      consumes:
      - application/json
      description: |-
        Approved members only, while the group is **open**. `paid_by` must equal the caller — you can only record what *you* paid.

        `amount` is integer **baisa** (1.000 OMR = 1000) and must be positive. `occurred_on` must fall inside the trip's date range. `category` is one of: lodging, fuel, food, transport, other.
      parameters:
      - description: group id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: amount (baisa), category, description, occurred_on, paid_by
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.RecordExpenseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: the recorded expense
          schema:
            $ref: '#/definitions/types.Expense'
        "400":
          description: bad amount/date/category, or paid_by is not the caller
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: not an approved member
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: group not found
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: group is closed
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Record an expense you paid
      tags:
      - expenses
  /groups/{id}/expenses/{expenseId}:
    delete:
      description: You may delete **only your own** expenses, and only while the group
        is open.
      parameters:
      - description: group id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: expense id
        format: uuid
        in: path
        name: expenseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: expense deleted
        "400":
          description: invalid group or expense id
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: not your expense
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: expense not found
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: group is closed
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Delete an expense
      tags:
      - expenses
    patch:
      consumes:
      - application/json
      description: You may edit **only your own** expenses, and only while the group
        is open. Any change to the amount is written to the audit trail with **before/after**
        values.
      parameters:
      - description: group id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: expense id
        format: uuid
        in: path
        name: expenseId
        required: true
        type: string
      - description: fields to change
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.UpdateExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: the updated expense
          schema:
            $ref: '#/definitions/types.Expense'
        "400":
          description: invalid ids or body
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: not your expense
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: expense not found
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: group is closed
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Update an expense
      tags:
      - expenses
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Bearer token from POST /auth/login. Format: `Bearer <access_token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...